refactor: migrate to new model modules in network clients
- Update imports from `zzupy.models` to `zzupy.model.network`
